+++ changelog.md
@@ -4,7 +4,7 @@
 
 - Bugfix in `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.from_causal_graph` where the method was
   not adding the node correctly.
-- Added properties `max_time_lag` and `min_time_lag` to `cai_causal_graph.time_series_causal_graph. TimeSeriesCausalGraph` to
+- Added `max_time_lag` and `min_time_lag` properties to `cai_causal_graph.time_series_causal_graph. TimeSeriesCausalGraph` to
   return the maximum and minimum lag of the graph, respectively. The property `maxlag` was removed to avoid confusion.
 
 ## 0.3.7