+++ changelog.md
@@ -6,7 +6,9 @@
   `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` where the method was not adding the node correctly.
 - Added `max_time_lag` and `min_time_lag` properties to
   `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` to return the maximum and minimum lag of the graph,
-  respectively. The property `maxlag` was removed to avoid confusion.
+  respectively.
+- Fixed a bug with the property `maxlag`: it was removed to avoid confusion since it could give wrong information
+  if the future was included in the graph.
 
 ## 0.3.7